Major update! Pipes will never clip again!
The reason why before there would be clipping was due to my lack of knowledge of unity. In Unity, a cylinder (of which I recolor and call a Pipe) always has a height of 2 on scale 1,1,1. However, since unity does not include information like height of a gameobject in the editor, I could not know. So whenever my alreadyFilled function would check an area to see if a pipe already exists there - it was looking at the wrong area. This is major!!! The project is almost done!

After this there will be testing on whether I can make the performance of AttachPipe better (might swap to a coroutine since its no longer recursive). Excited for the next update!