Minor optimization change
Changed checkSphere's size to .9 (1.8 total)  to help with optimization: and creating a sphere a size of .9 compared to 1.0 REALLY helps when called every second. 

Will try to look for other ways to optimize the code further, if not I will proceed and attempt to finish this project the by the weekend.