Pipe pooler part 1
The pipe pooler is here and already its ability to save memory and speed is staggering. There is a slight problem now that spheres are not included in the pooler (I will either extend functionality to the pooler to allow for there to be interchangable pipes or something else). Its admittedly somewhat funny since all that's left are a bunch of spheres. Either way, excited to continue.