Beautifully broken update!
RIght now this program is a worse state than it was before - but that's ok. Right now I learned the importance as to why a pooler should be a shared instance. I may continue with making it a shared instance pooler, or I may try to be stubborn and try to make it work indivisually with each script. Either way: I am excited!